Added docstrings and refactored declarations.
Some simple docstrings were added to the functions in assigner.py

I refactored the functions in assigner.py to have variable declarations at the topmost level that made sense.
The reason was that `args` is not very explicit as to what is available.
Having the variable declarations at the top at least gives you an idea of what you're working with.

Lastly, I converted the older % string formatting to `.format()` for consistency.