Switch status command to use PrettyTable for output printing
Note: This introduces a dependency on PrettyTable. A line was added for that in requirements.txt.
However, you need to install it for the virtualenv, and the current assigner bash script does not have
an option to do that. A temporary workaround would be to run the pip install command manually for now.